# api.yaml
openapi: 3.0.3
info:
  title: QueryDataLayer
  description: Data Layer managing information about past queries.
  version: 1.0.0
servers:
  - url: https://query-dl/api/v1
paths:
  /:
    get:
      summary: Get queries
      description: Get the queries applicable to the specified parameters
      operationId: getQueries
      parameters:
        - name: location
          in: query
          description: General position of the search
          required: true
          schema:
            $ref: '#/components/schemas/Area'
        - name: tags
          in: query
          description: Tags of the sitting spot
          required: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Tag'
        - name: labels
          in: query
          description: Labels of the sitting spot
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: Queries found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref : '#/components/schemas/Query'
    post:
      summary: Post a new query
      description: create a new query entry to be later updated with the result
      operationId: postQuery
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryInDTO'
        required: true
      responses:
        201:
          description: Query entry created
          content:
            application/json:
              schema:
                type: string
                format: uuid
        400:
          description: Invalid input
  /{Id}:
    put:
      parameters:
        - name: Id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      summary: Updates a query entry with the result
      description: Add the result of a query
      operationId: updateQuery
      requestBody:
        description: The sitting spot found with the relative query
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/QueryResult'
      responses:
        200:
          description: Query updated
        400:
          description: Invalid input
        404:
          description: Query not found
components:
  schemas:
    Location:
      description: x,y coordinates
      properties:
        x:
          type: number
        y:
          type: number
    Area:
      description: An area with a location as center
      properties:
        center:
          $ref: '#/components/schemas/Location'
        range:
          type: number
    Tag:
      properties:
        key:
          type: string
        value:
          type: string
    QueryInDTO:
      properties:
        area:
          $ref: '#/components/schemas/Area'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        labels:
          type: array
          items:
            type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/QueryResult'
    Query:
      properties:
        queryId:
          type: string
          format: uuid
        area:
          $ref : '#/components/schemas/Area'
        tags:
          type: array
          items:
            $ref : '#/components/schemas/Tag'
        labels:
          type: array
          items:
            type: string
        results:
          type: array
          items:
            $ref: '#/components/schemas/QueryResult'
    QueryResult:
      properties:
        spotId:
          type: string
        location:
          $ref : '#/components/schemas/Location'